% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rf_model.R
\name{fit_rf_model}
\alias{fit_rf_model}
\title{Random Forest Model generation}
\usage{
fit_rf_model(dat_file)
}
\arguments{
\item{dat_file}{}
}
\description{
Random Forest Model generation
}
\examples{

fit_rf_model()
}
