% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tree_model.R
\name{fit_tree_model}
\alias{fit_tree_model}
\title{Decision Tree Model build and plot the decision tree-saved in graphs.}
\usage{
fit_tree_model(dat_file = TRUE)
}
\description{
Decision Tree Model build and plot the decision tree-saved in graphs.
}
\examples{

fit_tree_model()
}
