% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master.R
\name{Master}
\alias{Master}
\title{A BMS Subscription Model generation}
\usage{
Master(dat_file = TRUE)
}
\arguments{
\item{dat_file}{}
}
\description{
A BMS Subscription Model generation
}
\examples{


Master()
predict the test data
 Extract AUC Value for Random Forest,Decision Tree. Logistic Regression
}
