% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_glm_model.R
\name{perf_glm_model}
\alias{perf_glm_model}
\title{Performance of GLM models on test data}
\usage{
perf_glm_model(glm_model, test_set)
}
\arguments{
\item{glm_model, test_set}{}
}
\description{
Performance of GLM models on test data
}
\examples{

perf_glm_model()
}
