% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_rf_model.R
\name{perf_rf_model}
\alias{perf_rf_model}
\title{Performance of RF models on test data
This is used to also generate the AUC values, get FPR, TPR and get the data
for plotting the ROC curve for the model.Alos, to get the cut-offs.}
\usage{
perf_rf_model(rf_model, test_set)
}
\arguments{
\item{rf_model, test_set}{}
}
\description{
Performance of RF models on test data
This is used to also generate the AUC values, get FPR, TPR and get the data
for plotting the ROC curve for the model.Alos, to get the cut-offs.
}
\examples{

perf_rf_model()
}
