% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_tree_model.R
\name{perf_tree_model}
\alias{perf_tree_model}
\title{Performance of Decision Tree Model on test data}
\usage{
perf_tree_model(tree_model, test_set)
}
\arguments{
\item{tree_model, test_set}{}
}
\description{
Performance of Decision Tree Model on test data
}
\examples{

perf_tree_model()
}
