% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eda.R
\name{plot_eda}
\alias{plot_eda}
\title{Plot EDA
This creates all the plots and saves them under "graphs"
which are used for initial exploratory data analysis.}
\usage{
plot_eda(train)
}
\arguments{
\item{train}{}
}
\description{
Plot EDA
This creates all the plots and saves them under "graphs"
which are used for initial exploratory data analysis.
}
\examples{


plot_eda(train)
}
