% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{savePlot}
\alias{savePlot}
\title{Save Plot
This function saves all plots in the path/PlotName provided.}
\usage{
savePlot(PlotName, myPlot)
}
\arguments{
\item{PlotName, myPlot}{}
}
\description{
Save Plot
This function saves all plots in the path/PlotName provided.
}
\examples{


savePlot(PlotName,myPlot)
}
