% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_rf_model.R
\name{test_rf_model}
\alias{test_rf_model}
\title{Test of RF models on sample data set}
\usage{
test_rf_model(rf_model, data_test)
}
\arguments{
\item{rf_model, data_test}{}
}
\description{
Test of RF models on sample data set
}
\examples{

test_rf_model()
}
