% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test.R
\name{train_test}
\alias{train_test}
\title{Data Preparation Step
Reads the source data file,
does the initial exploratory data analysis by creating
and saving plots on individual features}
\usage{
train_test(dat_file)
}
\arguments{
\item{dat_file}{}
}
\description{
Also creates the sample train and test data set after taking care of any factorizatons if any
}
\examples{


train_test(dat_file)
}
