% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rf_model1.R
\name{fit_rf_model1}
\alias{fit_rf_model1}
\title{Random Forest Model generation
This will generate the random forest model using caret library.}
\usage{
fit_rf_model1(dat_file = TRUE)
}
\arguments{
\item{dat_file}{}
}
\description{
Random Forest Model generation
This will generate the random forest model using caret library.
}
\examples{

fit_rf_model1()
}
