% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master.R
\name{Master}
\alias{Master}
\title{A BMS Subscription Model generation
It calls different functions for model generation
Performance testing on test-set
Plot ROC curves for all the different models to compare the performance
Models in consideration: Logistic Regression, Decision Tree, Random Forest
path_root should be used as setwd.}
\usage{
Master(path_root)
}
\arguments{
\item{path_root}{}
}
\description{
A BMS Subscription Model generation
It calls different functions for model generation
Performance testing on test-set
Plot ROC curves for all the different models to compare the performance
Models in consideration: Logistic Regression, Decision Tree, Random Forest
path_root should be used as setwd.
}
\examples{


Master()
}
