% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_tree_model.R
\name{perf_tree_model}
\alias{perf_tree_model}
\title{Performance of Decision Tree Model on test data
This is used to also generate the AUC values, get FPR, TPR and get the data
for plotting the ROC curve for the model.Also, to get the cut-offs.}
\usage{
perf_tree_model(tree_model, test_set)
}
\arguments{
\item{tree_model, test_set}{}
}
\description{
Performance of Decision Tree Model on test data
This is used to also generate the AUC values, get FPR, TPR and get the data
for plotting the ROC curve for the model.Also, to get the cut-offs.
}
\examples{

perf_tree_model()
}
